requirements.txt
python-dotenv
pdfplumber
google-genai
requests
rich
